Thoughts regarding suggested links for 'tutorials'
There are a great many Ember resources at this point, and the list we have here isn't a fair slice of those. A few of them are not currently up to date + some are _free_ and some are _paid_ - and in general, it seems too random. Are these suggested tutorials for CLI specific things? I think it would be hard to officially recommend Something like rock-and-roll vs mike's front-end-masters courses - let alone the other 12 or so resources / from $0 to $2k. I don't have a solution, but I wanted to bring it up. Either we build an extensive list to use across ember properties / or link to some outside list / or just leave it up to Google.

Either way, the official learning resources should be here. I'm not sure what the best URL is though: https://guides.emberjs.com/v2.14.0/getting-started/quick-start/  vs https://guides.emberjs.com/quick-start and https://guides.emberjs.com/tutorial (it would be great to have a 'latest' URL that was easy to remember)

Thoughts on that?

The only other suggestion is that tooling is moved up in the hierarchy.